requirements.txt
pytest
selenium
requests
pyyaml
allure-pytest
